llm/requirements.txt
fastapi==0.111.1
openai==1.37.1
requests==2.32.3
pillow==10.4.0
termcolor==2.4.0
python-decouple==3.8